==> R/QM12-02R-07-Brazas/02-ProcessImgAnaResY.R <==
# 02-ProcessImgAnaResY.R
# source('~/work/proj/QM12-02R-07-Brazas/R/02-ProcessImgAnaResY.R')
#
# This script processes a series of .csv files generated by the
# analySIS Imaging-C module 'AnaBseNozzle'. This extracts the
# COG-offset in the Y direction and compares it to the value
# for the jet straightness in the web direction read from
# an .RData file for the chip. The script prepares stacked
# plots for comparison and a linear model.

rm(list=ls())
library(rAnaLab)
library(xtable)

id <- c('qm-03727-1033LTAT-C2-A1',
        'qm-03728-1033LTAT-C2-C1',
        'qm-03729-1033LTAT-C3-A1',
        'qm-03730-1033LTAT-C3-A2',
        'qm-03731-1033LTAT-C3-C1')
chip <- c('1033LTAT-C2',
          '1033LTAT-C2',
          '1033LTAT-C3',
          '1033LTAT-C3',
          '1033LTAT-C3')

str.wd   <- "~/work/proj/QM12-02R-07-Brazas/R/"
i.digits <- 3
pdf.dir  <- '../TeX/pdf/'
png.dir  <- '../TeX/png/'
tab.dir  <- '../TeX/tab/'
txt.dir  <- '../out/'

# should not need to change below here...
samples <- data.frame(str.id=id, str.chip=chip)
n.samples <- nrow(samples)

# get the web-angle
getWebAngle <-function(x)
{
  theWebAngle <- js$web.angle[x+1]
}

# compute the standard error
stderr <- function(x) sqrt(var(x)/length(x))

# number the nozzles
nozzleNumber <- function(nozzleName, noz.per.seg=512)
{
  nozzle.number <- NA
  lA <- strsplit(nozzleName,"_")
  num.ss <- length(lA[[1]])
  if(num.ss==2)
  {
    seg <- toupper(lA[[1]][1])
    noz <- toupper(lA[[1]][2])
    lB <- strsplit(seg,"S")
    num.ssb <- length(lB[[1]])
    if(num.ssb==2)
    {
      offset <- -1
      seg.name <- (lB[[1]][2])
      if(seg.name=='A') offset=0
      if(seg.name=='B') offset=noz.per.seg
      if(seg.name=='C') offset=2*noz.per.seg
      if(seg.name=='D') offset=3*noz.per.seg
      if(seg.name=='E') offset=4*noz.per.seg
      if(offset > -1)
      {
        lC <- strsplit(noz,"N")
        num.ssc <- length(lC[[1]])
        if(num.ssc==2)
        {
          noz.num.in.seg <- as.numeric(lC[[1]][2])
          # numbers are zero based
          nozzle.number <- offset + noz.num.in.seg
        }
      }
    }
  } 
  nozzle.number
}

plot.comparison <- function(df)
{
  par(mfrow = c(2,1))
  oldpar <- par(mar=c(0.0,4.1,2.1,1.1))
  plot(df$nozzle.number,
       df$js.wa, pch=19, col='red',
       main=str.id, axes=F,
       xlab='Nozzle',
       ylab='JS Web Angle')
  axis(2)
  box()
  par(mar=c(3.1,4.1,0,1.1))
  plot(df$nozzle.number,
       df$dcogy.nm, pch=19, col='black',
       xlab='Nozzle',
       ylab='Delta COG Y [nm]')
  mtext('Nozzle', side=1, line=2)
  par(mfrow = c(1,1))
  par(oldpar)
}

setwd(str.wd)
dir.create(pdf.dir, showWarnings = F, recursive = T)
dir.create(png.dir, showWarnings = F, recursive = T)
dir.create(tab.dir, showWarnings = F, recursive = T)
dir.create(txt.dir, showWarnings = F, recursive = T)

for (i in 1:n.samples) {
  str.id   <- samples$str.id[i]
  str.chip <- samples$str.chip[i]
  
  str.sem.file <- paste('../dat/', str.id, '.csv', sep='')
  str.js.file  <- paste('../dat/', str.chip, '-JS.RData', sep='')
  data <- read.csv(str.sem.file, header=F, skip=1, as.is=T)
  load(str.js.file)
  data <- subset(data, data[, 3] > 0 & data[, 4] > 0)
  n.pts <- nrow(data)
  noz.name <- data[, 1]
  nozzle.number <- sapply(noz.name, nozzleNumber)
  num.min <- min(nozzle.number)
  num.max <- max(nozzle.number)
  str.range <- sprintf("%d nozzles within %d-%d", n.pts, num.min, num.max )
  id    <- round(data[, 3], i.digits) # in microns
  od    <- round(data[, 4], i.digits) # in microns
  dcogy <- round(1000.*data[, 8], i.digits) # in nm
  js.wa <- round(sapply(nozzle.number, getWebAngle), i.digits)
  df.good <- data.frame(nozzle.number=nozzle.number,
                        id.um=id,
                        od.um=od,
                        dcogy.nm=dcogy,
                        js.wa=js.wa)


  par(mfrow = c(1,1))
  
  the.fit <- ralScatterPlotLine(df.good$dcogy.nm, 
                                df.good$js.wa,
                                NULL,
                                str.x='delta COG-Y [nm]',
                                str.y='JS Web Angle',
                                str.title=str.id,
                                i.digits=3,
                                st.cex=1, a.cex=1,
                                lw.ax=2,
                                bReg=T,
                                bErrorBars=F)
  print(str.id)
  print(summary(the.fit))
  plot.comparison(df.good)
  
  str.pdf <- paste(pdf.dir, str.id,'-comp.pdf', sep='')
  pdf(file=str.pdf, width=9, height=6, pointsize=18)
  plot.comparison(df.good)
  dev.off()
  
  str.png <- paste(png.dir, str.id,'-comp.png', sep='')
  png(filename=str.png, width=1024, height=768, pointsize=28)
  plot.comparison(df.good)
  dev.off()

  str.pdf <- paste(pdf.dir, str.id,'-fit.pdf', sep='')

  pdf(file=str.pdf, width=9, height=6, pointsize=18)
  par(mfrow = c(1,1))
  the.fit <- ralScatterPlotLine(df.good$dcogy.nm, 
                                df.good$js.wa,
                                NULL,
                                str.x='delta COG-Y [nm]',
                                str.y='JS Web Angle',
                                str.title=str.id,
                                i.digits=3,
                                st.cex=1, a.cex=1,
                                lw.ax=2,
                                bReg=T,
                                bErrorBars=F)
  dev.off()

  str.png <- paste(png.dir, str.id,'-fit.png', sep='')
  png(filename=str.png, width=1024, height=768, pointsize=28)
  the.fit <- ralScatterPlotLine(df.good$dcogy.nm, 
                                df.good$js.wa,
                                NULL,
                                str.x='delta COG-Y [nm]',
                                str.y='JS Web Angle',
                                str.title=str.id,
                                i.digits=3,
                                st.cex=1, a.cex=1,
                                lw.ax=2,
                                bReg=T,
                                bErrorBars=F)

  dev.off()

  str.txt <- paste(txt.dir, str.id,'-fit.txt', sep='')
  sink(str.txt)
  print(str.id)
  print(summary(the.fit))
  sink()


  str.tex <- paste(tab.dir,str.id,'-corr.tex', sep='')
  str.label <- paste('tab:',str.id,sep='')
  str.caption <- paste('Correlation of SEM delta-COG-Y and jet
                     straightness web angle analysis from', str.id)
  xt <- xtable(the.fit,caption=str.caption,
               label=str.label, floating=F)

  print(xt)

  sink(str.tex)
  print(xt)
  sink()

  str.ei <- '\\endinput'
  cat(str.ei, file=str.tex, sep='\n', append=T)

}
